Fixed support links in templates
