+++ packages/app/src/cli/prompts/generate/extension.ts
@@ -2,7 +2,7 @@ import {extensionTypesGroups} from '../../constants.js'
 import {AppInterface} from '../../models/app/app.js'
 import {GenericSpecification} from '../../models/app/extensions.js'
 import {TemplateSpecification} from '../../api/graphql/template_specifications.js'
-import {getExtensionSpecificationsFromTemplate} from '../../services/generate/fetch-extension-specifications.js'
+import {getExtensionSpecificationsFromTemplate} from '../../models/extensions/templates.js'
 import {generateRandomNameForSubdirectory} from '@shopify/cli-kit/node/fs'
 import {renderSelectPrompt, renderTextPrompt} from '@shopify/cli-kit/node/ui'
 import {outputWarn} from '@shopify/cli-kit/node/output'
